requirements.txt
flask
python-dotenv
pywinrm
python-evtx